pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mofapipeline"
version = "0.1.0"
description = "Reusable, academically rigorous MOFA+ pipeline for multi-omics datasets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "scikit-learn>=1.3",
]

[project.optional-dependencies]
mofa = [
  "mofapy2",
  "mofax",
]

[project.scripts]
mofa-pipeline = "mofapipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src/python"}

[tool.setuptools.packages.find]
where = ["src/python"]
